requirements.txt
streamlit==1.33.0
torch==2.2.2
opencv-python==4.11.0.86
numpy==1.26.4
transformers==4.48.3
Pillow==10.2.0
gTTS==2.5.4
